requirements.txt
black==24.10.0
bump2version==1.0.1
coverage==7.6.2
pre_commit==4.0.1
pytest-cov==5.0.0
pytest==8.3.3
PyYAML==6.0.2
Sphinx==8.1.3
structlog==24.4.0
tox==4.21.2
twine==5.1.1
typer==0.12.5
types-PyYAML==6.0.12.20240917
